refactor images to use Images.js to render left and right images based on prop orientedLeft
